fix: split permission matrix dual-bootstrap into separate spec files
CI investigation revealed Jest environment teardown failures when two AppModule.compile() + app.init() calls happen in same Jest worker. TypeORM's DirectoryExportedClassesLoader lazy-loads entity metadata after Suite A teardown is scheduled, triggering 'ReferenceError: import after Jest environment torn down'.

Solution: Split into separate .spec.ts files, each with single bootstrap (matches existing workspace-rbac.e2e-spec.ts pattern).

Files added:
- workspaces-controller-reads.bootstrap.ts (shared setup/teardown)
- workspaces-controller-reads.register-matrix.ts (shared test registrations)
- workspaces-controller-reads.suite-a.permission-matrix.e2e-spec.ts (flag ON, strict)
- workspaces-controller-reads.suite-b.permission-matrix.e2e-spec.ts (flag unset, permissive)

Files deleted:
- workspaces-controller-reads.permission-matrix.e2e-spec.ts (dual-suite monolith)

Permission-matrix Jest config keeps clean shutdown (no forceExit); architectural target is explicit teardown, not forced process exit.

Test count unchanged: 75 total (8 harness + 3 guard-audit + 59 Suite A + 5 Suite B). Architectural claim of dual-state flag parity preserved; only test infrastructure changed.
